requirements.txt
google-genai==1.13.0
pyTelegramBotAPI==4.27.0
python-telegram==0.19.0
python-telegram-bot==22.0
asyncio==3.4.3
asyncpg==0.30.0
alembic==1.15.1
email_validator==2.2.0
python-dotenv==1.1.0